MAME Patches to SEAL 1.0.7
This patch must be applied over SEAL 1.0.7 to fix three problems:

1) Sound not working on some SoundBlaster PCI128 cards.
2) True 16-bit support (the unpatched library only uses the 8 most
significant bits of each sample and throws away the others, causing a
noticeable background noise).
3) Mono playback was fixed (helps PCI cards which can't emulate a SB
card capable of stereo playback).

Don't forget to recompile the library after applying the patch, and to
copy it to the djgpp/lib directory!